Conflicts:
	PROJECTNEWTRIAL/game.cpp

+++ PROJECTNEWTRIAL/rendermaze.h
@@ -13,7 +13,7 @@ class RenderMaze
 private:
     maze M;
     int width, height, scale;
-	int side; //Refers to width of each 3d object. 
+	double side; //Refers to width of each 3d object. 
 	bool use_3d;
 public:
 	bool is_done;
